refactor: sectionWithSecitonsをtypesから参照する
