+++ frontend/src/features/article/useArticle.ts
@@ -2,9 +2,8 @@ import { Article } from '@prisma/client'
 import { useCallback } from 'react'
 
 import { useGetApi } from '@/hooks/useApi'
+import { SectionWithRelation } from '@/types'
 import { postApi, putApi, deleteApi } from '@/utils/api'
-
-import { SectionWithRelation } from '../section'
 
 // sectionのmutate
 export const useCreateArticle = (sectionId: string) => {
