+++ frontend/src/features/section/index.ts
@@ -5,4 +5,3 @@ export * from './components/SectionItem'
 export * from './components/SectionArticles'
 export * from './components/SectionQuizzes'
 export * from './useSection'
-export * from './types'
